# [FEM.Elastic] Simplify addkToMatrix in TriangularFEMForceFieldOptim