@alxbilger 2022-04-15
> Just realized in the description, is it slower after?

The difference is not significative. The timer to focus in this PR is MBKBuild.